+++ Extensions/dnSpy.Debugger/dnSpy.Debugger.DotNet.CorDebug/Dialogs/AttachToProcess/DotNetFrameworkAttachProgramOptionsProvider.cs
@@ -54,7 +54,7 @@ namespace dnSpy.Debugger.DotNet.CorDebug.Dialogs.AttachToProcess {
 						sb.EnsureCapacity((int)chBuffer);
 						hr = rtInfo.GetVersionString(sb, ref chBuffer);
 
-						yield return new DotNetFrameworkAttachProgramOptions((uint)process.Id, sb.ToString());
+						yield return new DotNetFrameworkAttachProgramOptions(process.Id, sb.ToString());
 					}
 				}
 			}
@@ -62,7 +62,7 @@ namespace dnSpy.Debugger.DotNet.CorDebug.Dialogs.AttachToProcess {
 	}
 
 	sealed class DotNetFrameworkAttachProgramOptions : AttachProgramOptions {
-		public override ulong ProcessId { get; }
+		public override int ProcessId { get; }
 		public override RuntimeId RuntimeId { get; }
 		public override string RuntimeName { get; }
 		public override Guid RuntimeGuid => PredefinedDbgRuntimeGuids.DotNetFramework_Guid;
@@ -70,7 +70,7 @@ namespace dnSpy.Debugger.DotNet.CorDebug.Dialogs.AttachToProcess {
 
 		readonly string debuggeeVersion;
 
-		public DotNetFrameworkAttachProgramOptions(ulong pid, string clrVersion) {
+		public DotNetFrameworkAttachProgramOptions(int pid, string clrVersion) {
 			ProcessId = pid;
 			RuntimeId = new DotNetFrameworkRuntimeId(clrVersion);
 			RuntimeName = "CLR " + clrVersion;
