+++ Extensions/dnSpy.Debugger/dnSpy.Debugger/Attach/AttachableProcessImpl.cs
@@ -23,7 +23,7 @@ using dnSpy.Contracts.Debugger.Attach;
 
 namespace dnSpy.Debugger.Attach {
 	sealed class AttachableProcessImpl : AttachableProcess {
-		public override ulong ProcessId => attachableProcessInfo.ProcessId;
+		public override int ProcessId => attachableProcessInfo.ProcessId;
 		public override RuntimeId RuntimeId => attachableProcessInfo.RuntimeId;
 		public override Guid RuntimeGuid => attachableProcessInfo.RuntimeGuid;
 		public override Guid RuntimeKindGuid => attachableProcessInfo.RuntimeKindGuid;
