+++ Extensions/dnSpy.Debugger/dnSpy.Debugger/Breakpoints/Code/CondChecker/DbgCodeBreakpointFilterChecker.cs
@@ -42,7 +42,7 @@ namespace dnSpy.Debugger.Breakpoints.Code.CondChecker {
 
 		sealed class DbgFilterEEVariableProviderImpl : DbgFilterEEVariableProvider {
 			public override string MachineName => Environment.MachineName;
-			public override ulong ProcessId => (uint)process.Id;
+			public override int ProcessId => process.Id;
 			public override string ProcessName => process.Filename;
 			public override ulong ThreadId => thread?.Id ?? ulong.MaxValue;
 			public override string ThreadName => thread?.UIName;
